+++ src/main/java/org/intellij/sequencer/Constants.java
@@ -4,4 +4,5 @@ public interface Constants {
     String ANONYMOUS_CLASS_NAME = "Anonymous";
     String CONSTRUCTOR_METHOD_NAME = "new";
     String Lambda_Invoke = "λ→";
+    String TOP_LEVEL_FUN = "Global";
 }