+++ src/main/java/org/intellij/sequencer/generator/ClassDescription.java
@@ -49,4 +49,6 @@ public class ClassDescription {
     }
 
     public static ClassDescription ANONYMOUS_CLASS = new ClassDescription(Constants.ANONYMOUS_CLASS_NAME,new ArrayList<>());
+
+    public static ClassDescription TOP_LEVEL_FUN = new ClassDescription(Constants.TOP_LEVEL_FUN, new ArrayList<>());
 }