+++ src/main/java/org/intellij/sequencer/generator/IGenerator.java
@@ -1,0 +1,8 @@
+package org.intellij.sequencer.generator;
+
+import com.intellij.psi.PsiElement;
+import com.intellij.psi.PsiMethod;
+
+public interface IGenerator {
+    CallStack generate(PsiElement psiElement);
+}